Stop the recovery prompt restating the file the run render already names
`render_recovery_prompt` pushed `render_local_ignore` into the prompt's lines, and
`render_run_result` pushes the same line into the Installed YAML Data block. Every
surface that receives a prompt receives the rendered run on the same envelope, so
all three interactive frontends showed the malformed file's identity twice once
they started rendering the prompt -- the drift this crate exists to remove, merely
relocated into core.

The prompt now states the question and nothing else: why the run paused, plus a
notice naming any decision being withheld. The identity is stated once, by the run
render, which is where it was already correct. That reduces
`render_recovery_prompt` to a single `bool`, and the "silence is not a denial"
rule stays where it landed, in `render_local_ignore_recovery`.

Also moves the recovery tests out of `lib_tests.rs` into `recovery_tests.rs`,
declared from `recovery.rs` with the path attribute, per AGENTS.md rule 11. They
run as `recovery::tests::*` now. The corpus in `lib_tests.rs` still renders the
prompt, which is deliberate rather than a leftover duplicate: the wording is
pinned beside the module that writes it, while the corpus is what subjects the
decision descriptions to the crate-wide placeholder and grammatical-number
invariants that no per-module file can apply.
