+++ business-logic/classic-scan-presentation/src/lib_tests.rs
@@ -9,9 +9,8 @@
 //! No test here runs a scan. Every render function is pure over a borrowed contract value.
 
 use super::{
-    DisplayLine, DisplaySegment, DisplaySeverity, RecoveryDecisionDescription, render_event,
-    render_infrastructure_error, render_local_ignore_recovery, render_resume_error,
-    render_run_result,
+    DisplayLine, DisplaySegment, DisplaySeverity, render_event, render_infrastructure_error,
+    render_local_ignore_recovery, render_resume_error, render_run_result,
 };
 use crate::display::{
     BYTE, CONCURRENT_SCAN, CRASH_LOG, CountedNoun, LOG, SEARCHED_LOCATION, TARGETED_INPUT,
@@ -26,16 +25,14 @@ use classic_config_core::{
 };
 use classic_scanlog_core::scan_run::contract::{
     Event, InfrastructureError, InfrastructureErrorStage, InstalledYamlDataRunDiagnosticKind,
-    LocalIgnoreRecoveryDecision, LocalIgnoreResetConflictError,
-    LocalIgnoreResetDurabilityUnknownError, LocalIgnoreResetFailure, LocalIgnoreResetFailureStage,
-    LocalIgnoreResetRunData, LocalIgnoreRunState, LogDisposition, LogEvent, LogFailure,
-    LogFailureStage, LogResult, ResumeError, RunResult, RunStatus,
+    LocalIgnoreResetConflictError, LocalIgnoreResetDurabilityUnknownError, LocalIgnoreResetFailure,
+    LocalIgnoreResetFailureStage, LocalIgnoreResetRunData, LocalIgnoreRunState, LogDisposition,
+    LogEvent, LogFailure, LogFailureStage, LogResult, ResumeError, RunResult, RunStatus,
 };
 use classic_scanlog_core::{
     CrashLogScanDiscoveryResult, CrashLogScanDiscoverySource, CrashLogScanRejectedInput,
     ScanProgressPhase,
 };
-use classic_vocabulary::Vocabulary;
 use std::path::PathBuf;
 
 use DisplaySegment::{Count, Emphasis, Label, Path, Text};
@@ -1004,168 +1001,6 @@ fn per_event_progress_pins_selected_concurrency() {
     );
 }
 
-// -- Local Ignore recovery decision descriptions -----------------------------------------
-
-/// Verifies the prompt states why the run paused and names the file being decided about.
-#[test]
-fn recovery_prompt_pins_its_situation_lines() {
-    let identity = YamlDataContentIdentity::from_bytes(b"malformed");
-
-    let prompt = render_recovery_prompt(
-        Some((LocalIgnoreRunState::RecoveryRequired, &identity)),
-        true,
-    );
-
-    assert_eq!(
-        prompt.lines,
-        vec![
-            DisplayLine {
-                severity: Warning,
-                segments: vec![Text(
-                    "Your Local Ignore file is malformed, so this Crash Log Scan is paused until you choose how to continue."
-                )],
-            },
-            render_local_ignore(LocalIgnoreRunState::RecoveryRequired, &identity),
-        ]
-    );
-}
-
-/// Verifies both decisions are described, in the contract's own variant order.
-///
-/// The descriptions are the locked Display Content this ticket exists to produce: pinned
-/// once here, and identical in every frontend and on every binding surface.
-#[test]
-fn recovery_prompt_pins_both_decision_descriptions() {
-    let identity = YamlDataContentIdentity::from_bytes(b"malformed");
-
-    let prompt = render_recovery_prompt(
-        Some((LocalIgnoreRunState::RecoveryRequired, &identity)),
-        true,
-    );
-
-    assert_eq!(
-        prompt.decisions,
-        vec![
-            RecoveryDecisionDescription {
-                decision: LocalIgnoreRecoveryDecision::ProceedWithoutIgnore,
-                label: "Proceed Without Ignore",
-                description: vec![Text(
-                    "Scan now with an empty ignore list. Your malformed Local Ignore file is left exactly as it is, and this choice applies to this scan only."
-                )],
-                available: true,
-            },
-            RecoveryDecisionDescription {
-                decision: LocalIgnoreRecoveryDecision::ResetToDefault,
-                label: "Reset To Default",
-                description: vec![Text(
-                    "Back up your malformed Local Ignore file byte-exactly, replace it with the selected Main defaults, then scan."
-                )],
-                available: true,
-            },
-        ]
-    );
-}
-
-/// Verifies an unavailable reset stays listed, marked unavailable, and says why.
-///
-/// Listed rather than dropped on purpose. A frontend must be able to explain the absence it
-/// is about to create, and it can only do that if core tells it what it is withholding.
-#[test]
-fn recovery_prompt_marks_an_unavailable_reset_and_explains_it() {
-    let identity = YamlDataContentIdentity::from_bytes(b"malformed");
-
-    let prompt = render_recovery_prompt(
-        Some((LocalIgnoreRunState::RecoveryRequired, &identity)),
-        false,
-    );
-
-    assert_eq!(
-        prompt.lines.last(),
-        Some(&DisplayLine {
-            severity: Notice,
-            segments: vec![Text(
-                "Reset To Default is unavailable: the selected Main YAML Data retains no usable default Local Ignore to publish."
-            )],
-        })
-    );
-    let unavailable: Vec<LocalIgnoreRecoveryDecision> = prompt
-        .decisions
-        .iter()
-        .filter(|description| !description.available)
-        .map(|description| description.decision)
-        .collect();
-    assert_eq!(
-        unavailable,
-        vec![LocalIgnoreRecoveryDecision::ResetToDefault]
-    );
-}
-
-/// Verifies proceeding is never withdrawn, whatever the run reported about reset.
-///
-/// This is the asymmetry that makes availability a per-decision field rather than one flag
-/// beside the prompt: Proceed Without Ignore needs nothing from the installation, so nothing
-/// can make it unavailable.
-#[test]
-fn recovery_prompt_never_withdraws_proceeding() {
-    let identity = YamlDataContentIdentity::from_bytes(b"malformed");
-
-    for reset_available in [true, false] {
-        let prompt = render_recovery_prompt(
-            Some((LocalIgnoreRunState::RecoveryRequired, &identity)),
-            reset_available,
-        );
-        let proceed = prompt
-            .decisions
-            .iter()
-            .find(|description| {
-                description.decision == LocalIgnoreRecoveryDecision::ProceedWithoutIgnore
-            })
-            .expect("Proceed Without Ignore is always described");
-        assert!(proceed.available);
-    }
-}
-
-/// Verifies the prompt describes exactly the decisions the continuation contract accepts.
-///
-/// Backing out is deliberately not among them: it is spelled as the *absence* of a decision
-/// and reaches the contract through `abandon`, so a frontend's cancel affordance is its own.
-#[test]
-fn recovery_prompt_describes_every_contract_decision_and_no_others() {
-    let identity = YamlDataContentIdentity::from_bytes(b"malformed");
-
-    let prompt = render_recovery_prompt(
-        Some((LocalIgnoreRunState::RecoveryRequired, &identity)),
-        true,
-    );
-
-    let described: Vec<LocalIgnoreRecoveryDecision> = prompt
-        .decisions
-        .iter()
-        .map(|description| description.decision)
-        .collect();
-    assert_eq!(described, LocalIgnoreRecoveryDecision::VARIANTS.to_vec());
-    assert_eq!(LocalIgnoreRecoveryDecision::VARIANTS.len(), 2);
-}
-
-/// Verifies a run that reported no Installed YAML Data keeps every decision on offer.
-///
-/// A run that reported nothing has not reported a denial. Withdrawing an option on silence
-/// would regress the behaviour that shipped before the availability fact existed, and this
-/// is the one place that rule is now written — the three frontends each used to write it.
-#[test]
-fn recovery_prompt_treats_absent_installed_yaml_data_as_available() {
-    let prompt = render_local_ignore_recovery(None);
-
-    assert!(
-        prompt
-            .decisions
-            .iter()
-            .all(|description| description.available)
-    );
-    // Nothing is claimed about a file whose identity the run never reported.
-    assert_eq!(prompt.lines.len(), 1);
-}
-
 // -- crate-wide invariants --------------------------------------------------------------
 
 /// Renders every path this crate has, so the invariant tests below see all of its output.
@@ -1376,19 +1211,18 @@ fn rendered_corpus() -> Vec<DisplayLine> {
         ));
     }
 
-    // The Local Ignore recovery prompt, in all three shapes it takes. Its decision
-    // descriptions are segment lists rather than lines, so they are wrapped as lines here:
-    // the invariant tests below read segments, and prose reaching a user through a
-    // description must be held to the same rule as prose reaching one through a line.
+    // The Local Ignore recovery prompt, in both shapes it takes. Its decision descriptions
+    // are segment lists rather than lines, so they are wrapped as lines here: the invariant
+    // tests below read segments, and prose reaching a user through a description must be held
+    // to the same rule as prose reaching one through a line.
+    //
+    // The wording itself is pinned in `recovery_tests.rs`, beside the module that renders it.
+    // What this adds is coverage by the crate-wide invariants, which no per-module test file
+    // can give it — a description that grew a placeholder or a re-derived plural would escape
+    // every check in this crate if the prompt were absent from this corpus.
     let mut prompts = vec![render_local_ignore_recovery(None)];
     for reset_available in [true, false] {
-        prompts.push(render_recovery_prompt(
-            Some((
-                LocalIgnoreRunState::RecoveryRequired,
-                &YamlDataContentIdentity::from_bytes(b"malformed"),
-            )),
-            reset_available,
-        ));
+        prompts.push(render_recovery_prompt(reset_available));
     }
     for prompt in prompts {
         lines.extend(prompt.lines);
