+++ docs/implementation/scan_run_presentation_consolidation.md
@@ -367,8 +367,10 @@ Details worth recording:
 - **The Qt buttons now read `Proceed Without Ignore` and `Reset To Default`.** They previously read `Continue Without Ignore` and `Back Up && Reset to Default` — the spelling the Display Label doc comment singles out as not using the word *Proceed* at all. Button text is the Display Label now; only the escaping of a literal `&` is the dialog's.
 - **The Python CLI renders the prompt without prompting.** A paused run stays terminal there — `docs/CLASSIC_Python_CLI_PRD.md` puts interactive prompts in CI-oriented commands out of scope — but terminal is not unexplained. The prompt's lines and every *available* decision reach both the plain stream and `data.recoveryPrompt`. It claims no continuation, so no file is touched and there is nothing to abandon. The filter costs that CLI nothing itself; it is there because a consumer reading that payload to drive its own prompt would otherwise repeat the two native frontends' bug.
 - **The three sentences joined `core-owned-phrases.txt` with this change**, not before it, exactly as the entry below predicted. All five audits enforce them now.
+- **`recovery.rs` carries its tests in `recovery_tests.rs`**, per rule 11, rather than in `lib_tests.rs` beside the crate-wide invariants. The corpus in `lib_tests.rs` still renders the prompt, which is deliberate and not a duplicate: the wording is pinned once beside the module, while the corpus is what subjects the decision descriptions to the placeholder and grammatical-number invariants that no per-module file can apply. `display.rs` and `render.rs` still have no siblings; moving those is a larger change than this ticket.
 
 - **The list carries both decisions, and each carries its own availability.** The acceptance condition "only decisions core is willing to accept appear" is met by construction rather than by filtering: `decisions` is built by walking `LocalIgnoreRecoveryDecision::VARIANTS`, so it can neither offer something the contract will refuse nor omit something it accepts, and the two exhaustive `match`es behind it stop the crate compiling if a third variant is ever added. Filtering unavailable decisions out instead was rejected — a frontend must be able to explain the absence it is about to create, and it can only do that if core tells it what is being withheld.
+- **The prompt states the question, not the facts behind it.** `lines` originally re-rendered the malformed file's Local Ignore line, which put the same line on screen twice in all three interactive frontends — the drift this whole effort removes, merely relocated into core. It was dropped once the frontends landed and the duplication became visible: `render_run_result` already carries that line in the Installed YAML Data block, and every surface receiving a prompt receives the rendered run on the same envelope, so the identity is stated once rather than never. That reduced `render_recovery_prompt` to a single `bool` parameter.
 - **The unavailability sentence moved to core, and nothing else did.** `Reset To Default is unavailable: the selected Main YAML Data retains no usable default Local Ignore to publish.` was already written identically in the TUI overlay, the native CLI menu, and the Qt dialog, each with a comment saying core would own it once this renderer existed. It is a prompt *line* rather than part of the Reset decision's `description`, because a description says what the decision *does* and stays true whether or not this run can honor it.
 - **Proceed Without Ignore is hard-coded available.** It needs nothing from the installation, so no run can withdraw it. That the two answers differ is the argument for a per-decision field over one prompt-wide flag, made in code.
 - **No entry was added to `core-owned-phrases.txt` by the core phase.** Every phrase on that list must already be gone from every frontend — the file's own third condition — and at that point these three sentences were still written locally in four places. They joined the deny-list with the frontend adoption, not before it, or all five audits would have failed the moment the renderer landed.
